requirements.txt
flask==3.0.0
flask-socketio==5.3.6
python-dotenv==1.0.0
tabulate==0.9.0
requests==2.31.0
asyncio==3.4.3
websockets==12.0
python-engineio==4.8.0
python-socketio==5.10.0
gevent==23.9.1
gevent-websocket==0.10.1